IO archive write correct buffer byte count to file. (#90)

+++ public/IO/InputArchive.hpp
@@ -39,12 +39,11 @@ public:
 	void ImportBuffer(T data)
 	{
 		static_assert(std::is_pointer_v<T> && "Data buffer should be pointer.");
-		size_t bufferBytes;
-		m_pIStream->read(reinterpret_cast<char*>(&bufferBytes), sizeof(size_t));
+		uint64_t bufferBytes;
+		m_pIStream->read(reinterpret_cast<char*>(&bufferBytes), sizeof(uint64_t));
 		if constexpr (SwapBytesOrder)
 		{
-			// bufferBytes = std::byteswap(bufferBytes);
-			bufferBytes = byte_swap<size_t>(bufferBytes);
+			bufferBytes = byte_swap<uint64_t>(bufferBytes);
 		}
 		m_pIStream->read(reinterpret_cast<char*>(data), bufferBytes);
 	}
@@ -58,7 +57,6 @@ public:
 			m_pIStream->read(reinterpret_cast<char*>(&data), sizeof(data));
 			if constexpr (SwapBytesOrder)
 			{
-				// data = std::byteswap(data);
 				data = byte_swap<T>(data);
 			}
 		}
@@ -67,12 +65,10 @@ public:
 			m_pIStream->read(reinterpret_cast<char*>(&data), sizeof(data));
 			if constexpr (4 == sizeof(T))
 			{
-				// data = std::byteswap(static_cast<uint32_t>(data));
 				data = byte_swap<float>(data);
 			}
 			else if constexpr (8 == sizeof(T))
 			{
-				// data = std::byteswap(static_cast<uint64_t>(data));
 				data = byte_swap<double>(data);
 			}
 			else
@@ -82,12 +78,11 @@ public:
 		}
 		else if constexpr (std::is_same<T, std::string>())
 		{
-			size_t dataLength;
-			m_pIStream->read(reinterpret_cast<char*>(&dataLength), sizeof(size_t));
+			uint64_t dataLength;
+			m_pIStream->read(reinterpret_cast<char*>(&dataLength), sizeof(uint64_t));
 			if constexpr (SwapBytesOrder)
 			{
-				// dataLength = std::byteswap(dataLength);
-				dataLength = byte_swap<size_t>(dataLength);
+				dataLength = byte_swap<uint64_t>(dataLength);
 			}
 			data.resize(dataLength);
 			m_pIStream->read(reinterpret_cast<char*>(data.data()), dataLength);
